chore: Split bench into multiple files and rename to perf
